requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
flask-cors==4.0.1
flasgger==0.9.7.1
python-dotenv==1.0.1
gunicorn==22.0.0
requests==2.32.3
flask-jwt-extended==4.6.0
pyjwt==2.9.0
requests-oauthlib==1.3.1
requests>=2.31.0